requirements.txt
dgl_cu102==0.9.1.post1
joblib==1.1.0
joblib==1.0.1
numpy==1.18.0
pandas==1.2.4
torch==1.12.1
torch==1.4.0
